requirements.txt
cryptography==2.7
psycopg2-binary==2.8.3
pillow==6.1.0
numpy==1.16.4
scipy==1.3.0
pandas==0.24.2
scikit-learn==0.21.2
nltk==3.4.4
six==1.12.0
smart-open==1.8.4
watchdog==0.9.0
pdf2image==1.6.0
pytesseract==0.2.7
gensim==3.7.3
tensorflow==2.0.0b1
tfx==0.13.0
jupyter==1.0.0
celery==4.1.1
kombu==4.2.0
tornado==5.1.1
